@@ -4,7 +4,7 @@ build-backend = "maturin"
 
 [project]
 name = "maptide"
-requires-python = ">=3.7"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Rust",
     "Programming Language :: Python :: Implementation :: CPython",
